Remove deprecated argument
When running a project with this package as dependency, I'm getting the following warning message.

```
RemovedInDjango40Warning: The providing_args argument is deprecated. As it is purely documentational, it has no replacement. If you rely on this argument as documentation, you can move the text to a code comment or docstring.
```

I'm moving the argument to a comment instead so it will not pollute the logs when developing in other projects 😊