Remove system check for Django version
Rationale: Dependency definitions are the correct way to ensure that compatible
versions of libraries are used. It is important that maintainers can divert from
that, for testing, or if they see fit, or in cases such as a dependency being
released and the maintainer not having the time to update the package.

Django follows SemVer, so it is quite safe to assume that if 3.1 works, 3.2 and
3.3 will as well.

Closes #175